+++ web-app/src/app/app.module.ts
@@ -3,20 +3,60 @@ import { NgModule } from '@angular/core';
 
 import { UpgradeModule } from '@angular/upgrade/static';
 import { UIRouterUpgradeModule } from '@uirouter/angular-hybrid';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgxMatSelectSearchModule } from 'ngx-mat-select-search';
+import { ScrollingModule } from '@angular/cdk/scrolling';
+
+import { 
+  MatIcon,
+  MatButton,
+  MatToolbar,
+  MatSpinner,
+  MatFormField,
+  MatIconModule,
+  MatButtonModule,
+  MatChipsModule,
+  MatToolbarModule,
+  MatProgressSpinnerModule,
+  MatFormFieldModule,
+  MatInputModule,
+  MatAutocompleteModule,
+  MatSelectModule
+} from '@angular/material';
+
+import { SwaggerComponent } from './swagger/swagger.component';
+import { ScrollWrapperComponent } from './wrapper/scroll/feed-scroll.component';
+import { DropdownComponent } from './observation/edit/dropdown/dropdown.component';
+import { MultiSelectDropdownComponent } from './observation/edit/multiselectdropdown/multiselectdropdown.component';
 
 import app from '../ng1/app.js';
-import { SwaggerComponent } from './swagger/swagger.component';
 
 @NgModule({
-  declarations: [SwaggerComponent],
+  declarations: [SwaggerComponent, DropdownComponent, MultiSelectDropdownComponent, ScrollWrapperComponent],
   imports: [
     BrowserModule,
     UpgradeModule,
-    UIRouterUpgradeModule.forRoot({ states: [] })
+    UIRouterUpgradeModule.forRoot({ states: [] }),
+    FormsModule,
+    ReactiveFormsModule,
+    BrowserAnimationsModule,
+    MatToolbarModule,
+    MatIconModule,
+    MatButtonModule,
+    MatProgressSpinnerModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatAutocompleteModule,
+    MatSelectModule,
+    NgxMatSelectSearchModule,
+    MatChipsModule,
+    ScrollingModule
   ],
   providers: [],
   bootstrap: [],
-  entryComponents:[SwaggerComponent]
+  entryComponents:[MatIcon, MatButton, MatToolbar, MatSpinner, MatFormField, DropdownComponent, MultiSelectDropdownComponent, ScrollWrapperComponent, SwaggerComponent]
 })
 export class AppModule { 
   constructor(private upgrade: UpgradeModule) { }