+++ web-app/src/app/wrapper/scroll/feed-scroll.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'feed-scroll-wrapper',
+  templateUrl: './feed-scroll.component.html',
+  styleUrls: []
+})
+export class ScrollWrapperComponent {
+
+  constructor() { }
+
+}