+++ web-app/src/ng1/app.js
@@ -6,6 +6,19 @@ import fileUpload from './file-upload/file.upload.component';
 import fileBrowser from './file-upload/file.browser.component';
 import  '@uirouter/angular-hybrid';
 import { SwaggerComponent } from "../app/swagger/swagger.component";
+import { downgradeComponent } from '@angular/upgrade/static';
+
+import {
+  MatIcon,
+  MatButton,
+  MatToolbar,
+  MatSpinner,
+  MatFormField
+} from '@angular/material';
+
+import { ScrollWrapperComponent } from '../app/wrapper/scroll/feed-scroll.component';
+import { DropdownComponent } from '../../src/app/observation/edit/dropdown/dropdown.component';
+import { MultiSelectDropdownComponent } from '../../src/app/observation/edit/multiselectdropdown/multiselectdropdown.component';
 
 require('angular-minicolors');
 require('select2');
@@ -24,6 +37,14 @@ const app = angular.module('mage', [
 ]);
 
 app
+  .directive('matIcon', downgradeComponent({ component: MatIcon }))
+  .directive('matButton', downgradeComponent({ component: MatButton }))
+  .directive('matToolbar', downgradeComponent({ component: MatToolbar }))
+  .directive('matSpinner', downgradeComponent({ component: MatSpinner }))
+  .directive('matFormField', downgradeComponent({ component: MatFormField }))
+  .directive('feedScrollWrapper', downgradeComponent({ component: ScrollWrapperComponent }))
+  .directive('observationEditDropdown', downgradeComponent({ component: DropdownComponent }))
+  .directive('observationEditMultiselectdropdown', downgradeComponent({ component: MultiSelectDropdownComponent }))
   .component('filterPanel', require('./filter/filter'))
   .component('exportPanel', require('./export/export'))
   .component('eventFilter', require('./filter/event.filter.component'))
@@ -64,7 +85,6 @@ require('./material-components');
 config.$inject = ['$httpProvider', '$stateProvider', '$urlRouterProvider', '$urlServiceProvider', '$animateProvider'];
 
 function config($httpProvider, $stateProvider, $urlRouterProvider, $urlServiceProvider, $animateProvider) {
-  // $urlRouterProvider.deferIntercept();
   $urlServiceProvider.deferIntercept();
   
   $httpProvider.defaults.withCredentials = true;