+++ web-app/src/ng1/observation/form/delete-observation.controller.js
@@ -12,7 +12,7 @@ function DeleteObservationController($scope, $uibModalInstance, EventService, ob
 
   $scope.delete = function() {
     EventService.archiveObservation($scope.observation).then(function() {
-      $scope.$emit('observation:delete',  $scope.observation);
+      $uibModalInstance.close();
     });
   };
 }